requirements.txt
# LangChain/LangGraph Core
langchain>=0.3.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# MCP Integration
langchain-mcp-adapters>=0.1.0

# Telegram Bot
python-telegram-bot>=21.0
python-telegram-bot[callback-data]

# Multimodal Processing
pillow>=10.0.0
pydub>=0.25.1
opencv-python>=4.8.0
moviepy>=1.0.3

# Model Integration
transformers>=4.45.0
torch>=2.0.0
accelerate>=0.25.0
vllm>=0.6.0  # For efficient LLM serving

# HTTP & API
httpx>=0.25.0
aiohttp>=3.9.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
tenacity>=8.2.0
loguru>=0.7.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.10.0
ruff>=0.1.0
